Fix ros_entrypoint.sh overwrite (#1009)
* Update Dockerfile.ros2 COPY command to fix "exec error"

When building ros2, an exec error was occuring when the container was trying to run the ENTRYPOINT command, and for some reason this was tracked down to the ros_entrypoint.sh file being empty inside the container.

Changing the single COPY command to two commands (for ros_entrypoint and ros2_install) fixed it.

* Revert "Update Dockerfile.ros2 COPY command to fix "exec error""

This reverts commit 08bbc6e42817852bf627af3aa0bd139b1d1b2d4e.

* Fix ros_entrypoint.sh overwrite

Fixed a problem where ros_entrypoint.sh was empty when building stages after ros_humble-desktop (e.g. ros_humble-desktop). It appeared ros_humble-foxglove stage called the ros2_install.sh script, which made the ros_entrypoint.sh file blank and caused an exec error.

- Added checks in ros2_install.sh to prevent overwriting ros_entrypoint.sh if the file is missing, invalid or already modified.
- Ensured ros_entrypoint.sh is only modified if it contains the expected ros_source_env function.
- Automatically sets executable permissions on ros_entrypoint.sh after editing, preventing runtime exec errors.
- This improves reliability of multi-layer Docker builds where ros_entrypoint.sh is inherited and extended across container stages.