fix: tensorrt missing SYSTEM_X86 switch to determine compatible tensorrt/cuda mapping
Before this fix, on x86 the command `LSB_RELEASE=24.04 CUDA_VERSION=12.8 ros:jazzy-ros-base` resulted in a broken dependency tree where is was installing both CUDA 12.8 and CUDA 12.9 in addition to the two different respective versions of CUDNN.

Intuitively it makes sense that the compatability matrix on x86 would line up with jetson/arm, but worth double checking.

This fixed my issue and was able to successfully build, run, and pass tests for all intermediate packages for `LSB_RELEASE=24.04 CUDA_VERSION=12.8 ros:jazzy-ros-base` on x86 with nvidia-container-runtime.