Modify the Dockerfile of huggingface_hub to make the installation of dataclasses conditional (#1217)
The dataclasses package only needs to be installed as a backport in Python 3.6. Starting from Python 3.7, dataclasses became part of the standard library. Since you are using Python 3.10 in your environment, it is normal that pip cannot find this package.
Modify the Dockerfile of huggingface_hub to make the installation of dataclasses conditional